% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtsp.R
\name{fortify}
\alias{fortify}
\title{Arranges the \code{mtsp} solution for plotting.}
\usage{
fortify(x, dat, ...)
}
\arguments{
\item{x}{The list returned by \code{mtsp}.}

\item{dat}{A matrix of x,y positions passed to \code{mtsp}. Only needed when \code{return_all=FALSE} was used.}

\item{...}{Other arguments.}
}
\description{
Arranges the \code{mtsp} solution for plotting.
}
