% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtsp.R
\name{get_distances.mdmtspv_ga}
\alias{get_distances.mdmtspv_ga}
\title{Grabs the distances for a single salesperson}
\usage{
\method{get_distances}{mdmtspv_ga}(x, idx, paths, depot_mat, dmat, ...)
}
\arguments{
\item{x}{The list returned by \code{mtsp}.}

\item{idx}{An index representing a single list element of \code{best_tour}.}

\item{paths}{The list with the \code{best_tour} returned by \code{mtsp}.}

\item{depot_mat}{The matrix \code{D0} giving the distances between the depots and positions, returned by \code{mtsp}.}

\item{dmat}{The distance matrix for the positions.}

\item{...}{Not used.}
}
\value{
Returns the distance for a single salesperson.
}
\description{
Grabs the distances for a single salesperson
}
