% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdmtspv_ga2}
\alias{mdmtspv_ga2}
\title{Runs the MTSP solver}
\usage{
mdmtspv_ga2(
  xy,
  dmat,
  nSalesmen,
  CostType,
  popSize,
  numIter,
  Epsilon = 1e-10,
  return_all = FALSE
)
}
\arguments{
\item{xy}{A \code{n} by \code{2} matrix of coordinate positions. Each row has the \code{x} and \code{y} position of the n sites to be visited.}

\item{dmat}{A \code{n} by \code{n} matrix with the distances between the positions \code{xy}.}

\item{nSalesmen}{An integer giving the number of salespersons (must be >= 2).}

\item{CostType}{Possible values are the integers 1 or 2; 1 minimises the total distance travelled and 2 minimises the maximum distance travelled across the salepersons.}

\item{popSize}{The population size / search space of the GA (should be divisible by 8 but this is corrected within program).}

\item{numIter}{is the number of desired iterations for the algorithm to run after a new best solution is found.}

\item{Epsilon}{A double giving a multiplier on blancing the costs beteen total distance and minmax (only used for \code{costType} 2).}

\item{return_all}{Logical on whether ro return full ouput. Defaults to \code{FALSE}.}
}
\value{
Returns the solution to the MTSP. Depending on \code{return_all} may return the inputs and  \code{best_tour}: 
a matrix where each row gives the sites visited by each salesperson,  \code{minDist}: is the best cost found by the algorithm, 
and  \code{generation}: is the number of generations required by the algorithm to find the solution.
}
\description{
Runs the MTSP solver
}
\examples{
set.seed(1)
n = 25
nSalesmen = 5
xy = matrix(rnorm(n*2), ncol=2)
dmat = as.matrix(dist(xy))
mdmtspv_ga2(xy, dmat, nSalesmen, 1, 80, 10)
}
