% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mtsp_ga}
\alias{mtsp_ga}
\title{Runs the MTSP solver}
\usage{
mtsp_ga(
  xy,
  dmat,
  nSalesmen,
  minTour,
  CostType,
  popSize,
  numIter,
  Epsilon = 1e-10,
  return_all = FALSE
)
}
\arguments{
\item{xy}{A \code{n} by \code{2} matrix of coordinate positions. Each row has the \code{x} and \code{y} position of the n sites to be visited.}

\item{dmat}{A \code{n} by \code{n} matrix with the distances between the positions \code{xy}.}

\item{nSalesmen}{An integer giving the number of salespersons (must be >= 2).}

\item{minTour}{An integer giving the minimum tour length for any of the salesmen.}

\item{CostType}{Possible values are the integers 1 or 2; 1 minimises the total distance travelled and 2 minimises the maximum distance travelled across the salepersons.}

\item{popSize}{The population size / search space of the GA (should be divisible by 8 but this is corrected within program).}

\item{numIter}{is the desired number of iterations for the algorithm to run.}

\item{Epsilon}{A double giving a multiplier on blancing the costs beteen total distance and minmax (only used for \code{costType} 2).}

\item{return_all}{Logical on whether ro return full ouput. Defaults to \code{FALSE}.}
}
\value{
Returns the solution to the MTSP. Depending on \code{return_all} may return the inputs and \code{best_tour}: returns
 a list where each element gives the sites visited by each salesperson, \code{minDist}: is the best cost found by the algorithm,
 \code{minDistTotal} gives tht total distance travelled for all salespersons, and \code{minMaxTotal} gives the maximum distance travelled by any one salesperson.
}
\description{
Runs the MTSP solver
}
\examples{
set.seed(1)
n = 25
nSalesmen = 5
xy = matrix(rnorm(n*2), ncol=2)
dmat = as.matrix(dist(xy))
minTour = 3
numIter = 10
popSize = 8
CostType = 2
res = mtsp_ga(xy, dmat, nSalesmen, minTour, CostType, popSize, numIter, return_all = TRUE)
}
