% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtsp.R
\name{plot.mtsp_ga}
\alias{plot.mtsp_ga}
\alias{ggplotter}
\title{Plot the MTSP solution}
\usage{
\method{plot}{mtsp_ga}(x, dat = NULL, ...)

ggplotter(x, dat, ...)
}
\arguments{
\item{x}{The list returned by \code{mtsp}.}

\item{dat}{A matrix of x,y positions passed to \code{mtsp}. Only needed when \code{return_all=FALSE} was used.}

\item{...}{Unused.}
}
\description{
Plot the MTSP solution (currently only implemented for the non-depot algorithms).
}
\examples{
set.seed(1)
n = 25
xy = matrix(rnorm(n*2), ncol=2)
run = mtsp(xy, nSalesmen=5, CostType=2, popSize=80, 
           numIter=10, algorithm="mtsp_ga", return_all=TRUE)
plot(run) 
}
