% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtsp.R
\name{summary.mtsp}
\alias{summary.mtsp}
\title{Calculates the distance for each saleperson}
\usage{
\method{summary}{mtsp}(object, xy = NULL, ...)
}
\arguments{
\item{object}{The list returned by \code{mtsp}.}

\item{xy}{A \code{n} by \code{2} matrix of coordinate positions. Each row has the \code{x} and \code{y} position of the n sites to be visited.}

\item{...}{Not used.}
}
\value{
Returns the distance for each salesperson.
}
\description{
Calculates the distance for each saleperson
}
