% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtsp.R
\name{get_distances}
\alias{get_distances}
\title{Grabs the distances for a single salesperson}
\usage{
get_distances(x, ...)
}
\arguments{
\item{x}{The list returned by \code{mtsp}.}

\item{...}{Other arguments.}
}
\description{
Grabs the distances for a single salesperson
}
